Second commit

+++ 0x00-hello_world/6-size.c
@@ -11,7 +11,7 @@ int main(void)
 	char c;
 	int i;
 	long int li;
-	long int lli;
+	long long int lli;
 	float f;
 
 	printf("Size of a char: %lu byte(s)\n", sizeof(c));
